requirements.txt
requests
GitPython
pytrends
pandas